cumulative to be split into LCDW and LCCum
